Testing grok game 1 Prompt found on x
create a simple game that I can run locally in my browser
it should be in the style of the 1979 asteroids arcade classic
I want it to use only 5 additional files:
png for the hero
png for the enemy
png for weapon powerups
wav file for when I shoot an enemy
wav file for the looping background sound on lower volume
Title: Doge Defender
Score: display the score in billions of $ format with each enemy shot $10,000,000.
If you lose you see game over with r to restart
If you collect powerups your weapon adds an additional bullet for each shot
You win if you get to $1,000,000,000.